requirements.txt
# Dependências principais do projeto TimeFlow
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
mysqlclient>=2.2
pytz>=2024.1
weasyprint>=61.0
qrcode>=7.4
Pillow>=10.0

# Dependências de desenvolvimento (opcional)
# ipython>=8.0
# django-debug-toolbar>=4.0